Infer: server queue hanging bug and several other improvements

+++ src/first_move/model/client.py
@@ -1,11 +1,15 @@
 import uuid
 import zmq
+import io
+import sys
 import numpy as np
 from ..config import config
+import os
 
 class ModelClient:
     def __init__(self):
         self.context = zmq.Context()
+        #print("Process ID:", os.getpid(), self.context)
         self.socket = self.context.socket(zmq.DEALER)  # DEALER socket for request/reply
         self.client_id = uuid.uuid4().bytes  # Unique client ID
         self.socket.setsockopt(zmq.IDENTITY, self.client_id)  # Set client ID
@@ -24,7 +28,10 @@ class ModelClient:
         _, response = self.socket.recv_multipart()
         #response = self.socket.recv_multipart()
         #print(response)
-        pv = np.frombuffer(response, dtype=np.float32)  # Deserialize combined data
+        #arrays = np.load(io.BytesIO(response))
+        #policy = arrays['p']
+        #value = arrays['v']
+        pv = np.frombuffer(response, dtype=np.float16)  # Deserialize combined data
         policy = pv[:-1]  # First 4 elements are policy
         value = pv[-1]   # Last element is the scalar value
         print(policy, value)
